LUCENE-5944: backport changes to branch_5x, except in-flux changes (o.a.l.document change, removal of solr .war), plus merge changes entries, plus fix solr precommit, plus fix docs, plus keep standardtokenizer 4.0 grammar for back compat
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/branch_5x@1626055 13f79535-47bb-0310-9956-ffa450edef68

+++ solr/core/src/test/org/apache/solr/core/QueryResultKeyTest.java
@@ -39,7 +39,7 @@ public class QueryResultKeyTest extends SolrTestCaseJ4 {
     // the hashcode should be the same even when the list
     // of filters is in a different order
     
-    Sort sort = new Sort(new SortField("test", SortField.Type.BYTE));
+    Sort sort = new Sort(new SortField("test", SortField.Type.INT));
     BooleanQuery query = new BooleanQuery();
     query.add(new TermQuery(new Term("test", "field")), Occur.MUST);
